+++ processor/src/test/resources/expected/Arez_DefaultMethodsModel.java
@@ -66,7 +66,7 @@ public final class Arez_DefaultMethodsModel extends DefaultMethodsModel implemen
 
   final int $$arezi$$_id() {
     if ( Arez.shouldCheckInvariants() && !Arez.areNamesEnabled() && !Arez.areRegistriesEnabled() && !Arez.areNativeComponentsEnabled() ) {
-      Guards.fail( () -> "Method invoked to access id when id not expected." );
+      Guards.fail( () -> "Method invoked to access id when id not expected on component named '" + $$arezi$$_name() + "'." );
     }
     return this.$$arezi$$_id;
   }