+++ processor/src/test/resources/expected/Arez_SingletonWithIdModel.java
@@ -78,7 +78,7 @@ public final class Arez_SingletonWithIdModel extends SingletonWithIdModel implem
 
   final int $$arezi$$_id() {
     if ( Arez.shouldCheckInvariants() && !Arez.areNamesEnabled() && !Arez.areRegistriesEnabled() && !Arez.areNativeComponentsEnabled() ) {
-      Guards.fail( () -> "Method invoked to access id when id not expected." );
+      Guards.fail( () -> "Method invoked to access id when id not expected on component named '" + $$arezi$$_name() + "'." );
     }
     return this.$$arezi$$_id;
   }