+++ processor/src/test/resources/expected/NestedModel_Arez_BasicActionModel.java
@@ -55,7 +55,7 @@ public final class NestedModel_Arez_BasicActionModel extends NestedModel.BasicAc
 
   final int $$arezi$$_id() {
     if ( Arez.shouldCheckInvariants() && !Arez.areNamesEnabled() && !Arez.areRegistriesEnabled() && !Arez.areNativeComponentsEnabled() ) {
-      Guards.fail( () -> "Method invoked to access id when id not expected." );
+      Guards.fail( () -> "Method invoked to access id when id not expected on component named '" + $$arezi$$_name() + "'." );
     }
     return this.$$arezi$$_id;
   }