+++ processor/src/test/resources/expected/com/example/collections/Arez_ObservableNonnullListModel.java
@@ -67,7 +67,7 @@ public final class Arez_ObservableNonnullListModel extends ObservableNonnullList
 
   final int $$arezi$$_id() {
     if ( Arez.shouldCheckInvariants() && !Arez.areNamesEnabled() && !Arez.areRegistriesEnabled() && !Arez.areNativeComponentsEnabled() ) {
-      Guards.fail( () -> "Method invoked to access id when id not expected." );
+      Guards.fail( () -> "Method invoked to access id when id not expected on component named '" + $$arezi$$_name() + "'." );
     }
     return this.$$arezi$$_id;
   }