+++ processor/src/test/resources/expected/com/example/component_name_ref/Arez_ComponentTypeNameAloneOnSingletonModel.java
@@ -57,7 +57,7 @@ public final class Arez_ComponentTypeNameAloneOnSingletonModel extends Component
 
   final int $$arezi$$_id() {
     if ( Arez.shouldCheckInvariants() && !Arez.areRegistriesEnabled() && !Arez.areNativeComponentsEnabled() ) {
-      Guards.fail( () -> "Method invoked to access id when id not expected." );
+      Guards.fail( () -> "Method invoked to access id when id not expected on component named '" + $$arezi$$_name() + "'." );
     }
     return this.$$arezi$$_id;
   }