+++ processor/src/test/resources/expected/com/example/component_ref/Arez_AnnotatedComponent.java
@@ -77,7 +77,7 @@ final class Arez_AnnotatedComponent extends AnnotatedComponent implements Dispos
 
   final int $$arezi$$_id() {
     if ( Arez.shouldCheckInvariants() && !Arez.areNamesEnabled() && !Arez.areRegistriesEnabled() && !Arez.areNativeComponentsEnabled() ) {
-      Guards.fail( () -> "Method invoked to access id when id not expected." );
+      Guards.fail( () -> "Method invoked to access id when id not expected on component named '" + $$arezi$$_name() + "'." );
     }
     return this.$$arezi$$_id;
   }