+++ processor/src/test/resources/expected/com/example/dependency/ComponentDependencyModel_Arez_Foo.java
@@ -57,7 +57,7 @@ final class ComponentDependencyModel_Arez_Foo extends ComponentDependencyModel.F
 
   final int $$arezi$$_id() {
     if ( Arez.shouldCheckInvariants() && !Arez.areNamesEnabled() && !Arez.areRegistriesEnabled() && !Arez.areNativeComponentsEnabled() ) {
-      Guards.fail( () -> "Method invoked to access id when id not expected." );
+      Guards.fail( () -> "Method invoked to access id when id not expected on component named '" + $$arezi$$_name() + "'." );
     }
     return this.$$arezi$$_id;
   }