+++ processor/src/test/resources/expected/com/example/parameterized_type/Arez_ResolvedModel.java
@@ -63,7 +63,7 @@ final class Arez_ResolvedModel extends ResolvedModel implements Disposable, Iden
 
   final int $$arezi$$_id() {
     if ( Arez.shouldCheckInvariants() && !Arez.areNamesEnabled() && !Arez.areRegistriesEnabled() && !Arez.areNativeComponentsEnabled() ) {
-      Guards.fail( () -> "Method invoked to access id when id not expected." );
+      Guards.fail( () -> "Method invoked to access id when id not expected on component named '" + $$arezi$$_name() + "'." );
     }
     return this.$$arezi$$_id;
   }