+++ processor/src/test/resources/expected/com/example/repository/Arez_RepositoryWithDestroyAndDetachRepository.java
@@ -62,7 +62,7 @@ final class Arez_RepositoryWithDestroyAndDetachRepository extends RepositoryWith
 
   final int $$arezi$$_id() {
     if ( Arez.shouldCheckInvariants() && !Arez.areRegistriesEnabled() && !Arez.areNativeComponentsEnabled() ) {
-      Guards.fail( () -> "Method invoked to access id when id not expected." );
+      Guards.fail( () -> "Method invoked to access id when id not expected on component named '" + $$arezi$$_name() + "'." );
     }
     return this.$$arezi$$_id;
   }